Implement more widgets, start working on PDF viewer

+++ lv_conf.h
@@ -126,7 +126,7 @@
 /*===================
  *  LV_OBJ SETTINGS
  *==================*/
-#define LV_OBJ_FREE_NUM_TYPE    uint32_t    /*Type of free number attribute (comment out disable free number)*/
+#define LV_OBJ_FREE_NUM_TYPE    size_t      /*Type of free number attribute (comment out disable free number)*/
 #define LV_OBJ_FREE_PTR         1           /*Enable the free pointer attribute*/
 
 /*==================
